requirements.txt
pymongo>=4.6.0
pandas>=2.1.3
python-dotenv>=1.0.0
crewai>=0.34.3
google-generativeai>=0.3.0
flask